Reorganized webview structure (mongo refactoring)
