+++ src/webviews/documentdb/collectionView/collectionView.scss
@@ -5,7 +5,7 @@
 
 @import '../../theme/slickgrid.scss';
 
-@import '../mongoClustersStyles.scss';
+@import '../sharedStyles';
 
 //@import '@fluentui/react/dist/sass/_References';
