+++ src/webviews/documentdb/collectionView/components/MonacoAdaptive.tsx
