+++ src/webviews/documentdb/collectionView/components/QueryEditor.tsx
