+++ src/webviews/documentdb/collectionView/components/dataViewPanelTableV2.scss
