+++ src/webviews/documentdb/collectionView/components/monacoAdaptive.scss
