+++ src/webviews/documentdb/collectionView/components/toolbar/ToolbarDividerTransparent.tsx
