+++ src/webviews/documentdb/collectionView/components/toolbar/ViewSwitcher.tsx
