+++ src/webviews/documentdb/documentView/components/toolbarDocuments.tsx
