+++ src/webviews/documentdb/documentView/documentView.scss
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-@import '../mongoClustersStyles.scss';
+@import '../sharedStyles';
 
 .documentView {
     display: flex;
