+++ src/webviews/documentdb/sharedStyles.scss
